+++ src/api/grpc/clients.grpc.api.ts
@@ -1,4 +1,5 @@
 import { ApplyService } from "@ak-proto-ts/apply/v1/svc_connect";
+import { BuildsService } from "@ak-proto-ts/builds/v1/svc_connect";
 import { DeploymentsService } from "@ak-proto-ts/deployments/v1/svc_connect";
 import { EnvsService } from "@ak-proto-ts/envs/v1/svc_connect";
 import { ProjectsService } from "@ak-proto-ts/projects/v1/svc_connect";
@@ -11,3 +12,4 @@ export const environmentsClient = createPromiseClient(EnvsService, grpcTransport
 export const deploymentsClient = createPromiseClient(DeploymentsService, grpcTransport);
 export const sessionsClient = createPromiseClient(SessionsService, grpcTransport);
 export const manifestApplyClient = createPromiseClient(ApplyService, grpcTransport);
+export const buildsClient = createPromiseClient(BuildsService, grpcTransport);