+++ src/enums/index.ts
@@ -1,6 +1,7 @@
 export { Theme } from "@enums/theme.enum";
 export { MessageType } from "@enums/messageType.enum";
-export { SessionStateType } from "@enums/sessionStates.enum";
+export { SessionLogRecordType, SessionStateType } from "@enums/sessionStates.enum";
 export { LoggerLevel } from "@enums/output.enum";
 export { SortOrder } from "@enums/sortOrder.enum";
 export { ProjectIntervalTypes } from "@enums/projectIntervals.enum";
+export { DeploymentState } from "@enums/deploymentState.enum";