+++ src/models/views/deploymentsView.model.ts
@@ -2,6 +2,5 @@ import { Deployment } from "@type/models";
 
 export type DeploymentSectionViewModel = {
 	deployments?: Deployment[];
-	totalDeployments?: number;
-	selectedDeploymentId?: string;
+	totalDeployments: number;
 };