+++ src/services/projects.service.ts
@@ -1,8 +1,7 @@
-import { SetResourcesResponse } from "@ak-proto-ts/projects/v1/svc_pb";
 import { projectsClient } from "@api/grpc/clients.grpc.api";
 import { namespaces } from "@constants";
 import { translate } from "@i18n";
-import { convertProjectProtoToModel } from "@models";
+import { convertErrorProtoToModel, convertProjectProtoToModel } from "@models";
 import { DeploymentsService, EnvironmentsService, LoggerService } from "@services";
 import { ServiceResponse } from "@type";
 import { Project } from "@type/models";
@@ -34,20 +33,25 @@ export class ProjectsService {
 	}
 
 	static async build(projectId: string, resources: Record<string, Uint8Array>): Promise<ServiceResponse<string>> {
-		const { error: resourcesError } = await this.setResources(projectId, resources);
-		if (resourcesError) {
-			LoggerService.error(`${namespaces.projectService} - Upload resources`, (resourcesError as Error).message);
+		try {
+			await projectsClient.setResources({
+				projectId,
+				resources,
+			});
+			const { buildId, error } = await projectsClient.build({ projectId });
+			if (error) {
+				LoggerService.error(
+					`${namespaces.projectService} - Build: `,
+					convertErrorProtoToModel(error.value, projectId).message
+				);
 
-			return { data: undefined, error: resourcesError };
+				return { data: undefined, error };
+			}
+			return { data: buildId, error: undefined };
+		} catch (error) {
+			LoggerService.error(namespaces.projectService, `Project ID: ${projectId}, error:${(error as Error).message}`);
+			return { data: undefined, error: (error as Error).message };
 		}
-
-		const { buildId, error } = await projectsClient.build({ projectId });
-		if (error) {
-			LoggerService.error(`${namespaces.projectService} - Build`, error.message);
-
-			return { data: undefined, error };
-		}
-		return { data: buildId, error: undefined };
 	}
 
 	static async deploy(projectId: string, buildId: string): Promise<ServiceResponse<string>> {
@@ -98,21 +102,5 @@ export class ProjectsService {
 			return { data: undefined, error: activateError };
 		}
 		return { data: deploymentId, error: undefined };
-	}
-
-	static async setResources(
-		projectId: string,
-		resources: Record<string, Uint8Array>
-	): Promise<ServiceResponse<SetResourcesResponse>> {
-		try {
-			await projectsClient.setResources({
-				projectId,
-				resources,
-			});
-			return { data: undefined, error: undefined };
-		} catch (error) {
-			LoggerService.error(namespaces.resourcesService, (error as Error).message);
-			return { data: undefined, error };
-		}
 	}
 }