+++ vscode-react/src/hooks/useCloseOnEscape.tsx
@@ -1,0 +1,14 @@
+import { useEffect } from "react";
+
+export const useCloseOnEscape = (closeEvent: () => void) => {
+	useEffect(() => {
+		const handleKeyDown = (event: KeyboardEvent) => {
+			if (event.key === "Escape") {
+				closeEvent();
+			}
+		};
+
+		window.addEventListener("keydown", handleKeyDown);
+		return () => window.removeEventListener("keydown", handleKeyDown);
+	}, [closeEvent]);
+};